+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - spring-like conf reference
 - optional casting in configuration reference
 - For VertxModuleTest, added new deployment method which allows supplying custom set of config stores
+- consul-sd-registrar support IP/hostname discovery
 
 ### Fixed
 - sd-provider/consul and sd-provider/static modules deployment race condition